implement user registration with avatar upload and cloud storage integration

+++ src/utils/cloudinary.js
@@ -4,7 +4,7 @@ import fs from "fs";
 cloudinary.config({ 
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
   api_key: process.env.CLOUDINARY_API_KEY, 
-  api_secret: process.env.CLOUDINARY_API_SECRET
+  api_secret: process.env.CLOUDINARY_API_SECRET 
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
@@ -12,16 +12,20 @@ const uploadOnCloudinary = async (localFilePath) => {
     if(!localFilePath) return null
 
     const response = await cloudinary.uploader.upload(localFilePath, {
-      resource_type: "image",
-      folder: "todos-users-avatar",
-      allowed_formats: ["jpg", "png", "jpeg"]
+      resource_type: "auto",
+      folder: "todo-app/avatars"
     })
 
     fs.unlinkSync(localFilePath)
     return response;
 
   } catch (error) {
-    fs.unlinkSync(localFilePath)
+     console.error("Cloudinary Upload Error:", error);
+    
+    if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
+    }
+
     return null;
   }
 }
